+++ send_control.cpp
@@ -142,7 +142,7 @@ tint    Channel::CwndRateNextSendTime () {
         return SwitchSendControl(KEEP_ALIVE_CONTROL);
     if (data_out_.size()<cwnd_) {
         dprintf("%s #%u sendctrl next in %llius (cwnd %.2f, data_out %i)\n",
-                tintstr(),id_,send_interval_,cwnd_,(int)data_out_.size());
+                tintstr(),(unsigned int)id_,(long long int)send_interval_,(float)cwnd_,(int)data_out_.size());
         return last_data_out_time_ + send_interval_;
     } else {
         assert(data_out_.front().time!=TINT_NEVER);
@@ -223,7 +223,7 @@ tint Channel::LedbatNextSendTime () {
     }
 
     dprintf("%s #%u sendctrl ledbat %lli-%lli => %3.2f\n",
-            tintstr(),id_,owd_cur,owd_min,cwnd_);
+            tintstr(),(unsigned int)id_,(long long int)owd_cur,(long long int)owd_min,(float)cwnd_);
     return CwndRateNextSendTime();
 }
