+++ statsgw.cpp
@@ -78,7 +78,7 @@ static void StatsGwNewRequestCallback (struct evhttp_request *evreq, void *arg);
 void StatsExitCallback(struct evhttp_request *evreq)
 {
 	char contlenstr[1024];
-	sprintf(contlenstr,"%i",strlen(exit_page));
+	sprintf(contlenstr,"%i",(int)strlen(exit_page));
 	struct evkeyvalq *headers = evhttp_request_get_output_headers(evreq);
 	evhttp_add_header(headers, "Connection", "close" );
 	evhttp_add_header(headers, "Content-Type", "text/html" );
@@ -146,7 +146,7 @@ void StatsOverviewCallback(struct evhttp_request *evreq)
 	strcat(bodystr,bottom_page);
 
 	char contlenstr[1024];
-	sprintf(contlenstr,"%i",strlen(bodystr));
+	sprintf(contlenstr,"%i",(int)strlen(bodystr));
 	struct evkeyvalq *headers = evhttp_request_get_output_headers(evreq);
 	evhttp_add_header(headers, "Connection", "close" );
 	evhttp_add_header(headers, "Content-Type", "text/html" );
@@ -208,7 +208,7 @@ void StatsGetSpeedCallback(struct evhttp_request *evreq)
 	sprintf(speedstr,"{\"downspeed\": %d, \"success\": \"true\", \"upspeed\": %d, \"cdownspeed\": %d, \"cupspeed\": %d, \"nleech\": %d, \"nseed\": %d}", dspeed, uspeed, cdownspeed, cupspeed, nleech, nseed );
 
 	char contlenstr[1024];
-	sprintf(contlenstr,"%i",strlen(speedstr));
+	sprintf(contlenstr,"%i",(int)strlen(speedstr));
 	struct evkeyvalq *headers = evhttp_request_get_output_headers(evreq);
 	evhttp_add_header(headers, "Connection", "close" );
 	evhttp_add_header(headers, "Content-Type", "application/json" );
